+++ lib/samples/sliders/vertical_slider/customization/shape_customization/vertical_slider_tick_customization.dart
@@ -46,7 +46,7 @@ class _VerticalTickCustomizedSliderState extends SampleViewState {
         value: _value,
         onChanged: (dynamic values) {
           setState(() {
-            _value = values;
+            _value = values as double;
           });
         },
         interval: 10,
